[ISSUE-212] Package name refactoring with common rules
