+++ models/spring-ai-transformers/src/test/java/org/springframework/ai/embedding/ResourceCacheServiceTests.java
@@ -24,6 +24,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
+import org.springframework.ai.sbert.embedding.ResourceCacheService;
 import org.springframework.core.io.DefaultResourceLoader;
 
 import static org.assertj.core.api.Assertions.assertThat;
