+++ models/spring-ai-transformers/src/test/java/org/springframework/ai/embedding/TransformersEmbeddingClientTests.java
@@ -22,6 +22,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.ai.document.Document;
+import org.springframework.ai.sbert.embedding.TransformersEmbeddingClient;
 
 import static org.assertj.core.api.Assertions.assertThat;
